requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv==1.0.0

langchain==0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.10